wip: move spec list into runner-ct
