+++ npm/design-system/package.json
@@ -49,6 +49,7 @@
     "react": "16.8.6",
     "react-dom": "16.8.6",
     "rollup": "^2.38.5",
+    "rollup-plugin-copy-assets": "2.0.3",
     "rollup-plugin-peer-deps-external": "2.2.4",
     "rollup-plugin-postcss-modules": "2.0.2",
     "rollup-plugin-typescript2": "^0.29.0",
