+++ npm/design-system/rollup.config.js
@@ -6,6 +6,7 @@ import peerDepsExternal from 'rollup-plugin-peer-deps-external'
 import postcss from 'rollup-plugin-postcss-modules'
 import pkg from './package.json'
 import image from '@rollup/plugin-image'
+import copy from 'rollup-plugin-copy-assets'
 
 const banner = `
 /**
@@ -29,7 +30,17 @@ function createEntry (options) {
       'react-dom',
     ],
     plugins: [
-      peerDepsExternal(), resolve(), json(), commonjs(), postcss({ writeDefinitions: false }), image(),
+      peerDepsExternal(),
+      resolve(),
+      json(),
+      commonjs(),
+      postcss({ writeDefinitions: false }),
+      image(),
+      copy({
+        assets: [
+          './index.scss',
+        ],
+      }),
     ],
     output: {
       banner,
