+++ npm/design-system/src/index.ts
@@ -2,9 +2,9 @@ export * from './components/Button'
 
 export * from './components/CypressLogo/CypressLogo'
 
-export * from './components/FileExplorer/FileExplorer'
+// export * from './components/FileExplorer/FileExplorer'
 
-export * from './components/SpecList/SpecList'
+// export * from './components/SpecList/SpecList'
 
 export * from './components/SearchInput/SearchInput'
